requirements.txt
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.5
cycler==0.10.0
Django==3.1.5
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-taggit==1.3.0
Faker==5.6.5
flake8==3.8.4
future==0.18.2
html5lib==1.1
isort==5.7.0
jedi==0.17.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.4.1
mccabe==0.6.1
numpy==1.20.2
pandas==1.2.4
parso==0.7.1
Pillow==8.1.0
pluggy==0.13.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.5
reportlab==3.5.67
rope==0.18.0
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
ujson==4.0.1
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
yapf==0.30.0
